DELETE API 추가

+++ src/main/java/com/kuit/baemin/dto/request/ReviewReq.java
@@ -20,7 +20,7 @@ public class ReviewReq {
     @NotNull(message = "평점은 필수입니다.")
     @Min(value = 1, message = "평점은 최소 1점 이상이어야 합니다.")
     @Max(value = 5, message = "평점은 최대 5점 이하여야 합니다.")
-    private Integer rate; // 400 Bad Request 및 범위 검증용
+    private Double rate; // 400 Bad Request 및 범위 검증용
 
     @Size(max = 255, message = "리뷰 내용은 255자 이하로 작성해주세요.")
     private String comment;
